+++ service/grails-app/controllers/org/olf/UrlMappings.groovy
@@ -35,6 +35,8 @@ class UrlMappings {
 
     "/servint/widgets/types" (resources: 'widgetType')
 
+    "/erm/admin/$action"(controller:'admin')
+
     "/dashboard/definitions" (resources: 'widgetDefinition', method: 'GET')
   }
 }
